add SyntaxHighlighter
