+++ source-code/website/src/services/markdown/components/SyntaxHighlight.tsx
@@ -1,0 +1,10 @@
+import shiki from "shiki";
+
+const highlighter = await shiki.getHighlighter({ theme: "dark-plus" });
+
+export function SyntaxHighlight(props: { language?: string; content: string }) {
+	const code = highlighter.codeToHtml(props.content, {
+		lang: props.language,
+	});
+	return <div innerHTML={code}></div>;
+}
